Refactor Parser Tests for AddCommand
LogicManagerTest is too ambitious. It contains 500+ lines of code at the
moment and tests all commands in integration test level.

Let’s break it down to command-specific tests. This step moves out tests
pertaining to parsing of AddCommand.
